requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-allauth==0.45.0
django-boost==1.7.2
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.2
lxml==4.6.3
mysqlclient==2.0.3
oauthlib==3.1.1
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
soupsieve==2.2.1
sqlparse==0.4.1
ua-parser==0.10.0
urllib3==1.26.6
user-agents==2.2.0
whitenoise==5.3.0
